# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"MatTag Server","description":"API for material fingerprinting and analysis","version":"0.7.0"},"paths":{"/materials":{"get":{"tags":["Materials"],"summary":"Get Materials","operationId":"get_materials_materials_get","parameters":[{"name":"name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"}},{"name":"categories","in":"query","required":false,"schema":{"anyOf":[{"type":"array","items":{"$ref":"#/components/schemas/MaterialCategory"}},{"type":"null"}],"title":"Categories"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MaterialResponse"},"title":"Response Get Materials Materials Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Materials"],"summary":"Analyse Material","operationId":"analyse_material_materials_post","requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_analyse_material_materials_post"}}}},"responses":{"201":{"description":"Material analysis successful, data stored in database (store_in_db=True)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MaterialResponse"}}}},"200":{"description":"Material analysis successful, data NOT stored in database (store_in_db=False)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MaterialResponse"}}}},"400":{"description":"Bad request - invalid material name or image format"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/materials/{material_id}/image/specular":{"get":{"tags":["Materials"],"summary":"Get Material Specular Image","operationId":"get_material_specular_image_materials__material_id__image_specular_get","parameters":[{"name":"material_id","in":"path","required":true,"schema":{"type":"integer","title":"Material Id"}}],"responses":{"200":{"description":"Returns the specular image of the material as JPEG","content":{"image/jpeg":{}}},"404":{"description":"Image not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/materials/{material_id}/similar":{"get":{"tags":["Materials"],"summary":"Get Similar Materials","operationId":"get_similar_materials_materials__material_id__similar_get","parameters":[{"name":"material_id","in":"path","required":true,"schema":{"type":"integer","title":"Material Id"}},{"name":"name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"}},{"name":"categories","in":"query","required":false,"schema":{"anyOf":[{"type":"array","items":{"$ref":"#/components/schemas/MaterialCategory"}},{"type":"null"}],"title":"Categories"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MaterialResponse"},"title":"Response Get Similar Materials Materials  Material Id  Similar Get"}}}},"404":{"description":"Material with specified ID not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/materials/similar":{"post":{"tags":["Materials"],"summary":"Get Similar Materials By Characteristics","operationId":"get_similar_materials_by_characteristics_materials_similar_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SimilarMaterialsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/MaterialResponse"},"type":"array","title":"Response Get Similar Materials By Characteristics Materials Similar Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_analyse_material_materials_post":{"properties":{"specular_image":{"type":"string","format":"binary","title":"Specular Image","description":"Specular image of the material (JPEG or PNG)"},"non_specular_image":{"type":"string","format":"binary","title":"Non Specular Image","description":"Non specular image of the material (JPEG or PNG)"},"name":{"type":"string","title":"Name"},"category":{"$ref":"#/components/schemas/MaterialCategory"},"store_in_db":{"type":"boolean","title":"Store In Db"}},"type":"object","required":["specular_image","non_specular_image","name","category","store_in_db"],"title":"Body_analyse_material_materials_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MaterialCategory":{"type":"string","enum":["FABRIC","LEATHER","WOOD","METAL","PLASTIC","PAPER","COATING","UNCATEGORIZED"],"title":"MaterialCategory"},"MaterialCharacteristics":{"properties":{"brightness":{"type":"number","title":"Brightness"},"color_vibrancy":{"type":"number","title":"Color Vibrancy"},"hardness":{"type":"number","title":"Hardness"},"checkered_pattern":{"type":"number","title":"Checkered Pattern"},"movement_effect":{"type":"number","title":"Movement Effect"},"multicolored":{"type":"number","title":"Multicolored"},"naturalness":{"type":"number","title":"Naturalness"},"pattern_complexity":{"type":"number","title":"Pattern Complexity"},"scale_of_pattern":{"type":"number","title":"Scale Of Pattern"},"shininess":{"type":"number","title":"Shininess"},"sparkle":{"type":"number","title":"Sparkle"},"striped_pattern":{"type":"number","title":"Striped Pattern"},"surface_roughness":{"type":"number","title":"Surface Roughness"},"thickness":{"type":"number","title":"Thickness"},"value":{"type":"number","title":"Value"},"warmth":{"type":"number","title":"Warmth"}},"type":"object","required":["brightness","color_vibrancy","hardness","checkered_pattern","movement_effect","multicolored","naturalness","pattern_complexity","scale_of_pattern","shininess","sparkle","striped_pattern","surface_roughness","thickness","value","warmth"],"title":"MaterialCharacteristics"},"MaterialResponse":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"category":{"$ref":"#/components/schemas/MaterialCategory"},"characteristics":{"$ref":"#/components/schemas/MaterialCharacteristics"}},"type":"object","required":["id","name","category","characteristics"],"title":"MaterialResponse"},"SimilarMaterialsRequest":{"properties":{"characteristics":{"$ref":"#/components/schemas/MaterialCharacteristics"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"categories":{"anyOf":[{"items":{"$ref":"#/components/schemas/MaterialCategory"},"type":"array"},{"type":"null"}],"title":"Categories"}},"type":"object","required":["characteristics"],"title":"SimilarMaterialsRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}